requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
av==15.0.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blis==1.3.0
boto3==1.40.8
botocore==1.40.8
build==1.3.0
CacheControl==0.12.14
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.16
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.5
confluent-kafka==2.11.0
cryptography==45.0.6
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
dask==2022.10.2
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.116.1
faster-whisper==1.2.0
fbmessenger==6.0.0
filelock==3.18.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.2
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==2.12.0
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.14
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==5.2.0
murmurhash==1.0.13
networkx==2.6.3
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
orjson==3.11.2
overrides==7.7.0
packaging==21.3
pamqp==3.2.1
pandas==2.3.1
partd==1.4.2
pdf2image==1.17.0
pillow==11.3.0
pluggy==1.6.0
portalocker==2.10.1
posthog==5.4.0
preshed==3.0.10
prompt-toolkit==3.0.28
propcache==0.3.2
protobuf==4.23.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.33.2
pydot==1.4.2
Pygments==2.19.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.3
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
pytesseract==0.3.13
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2022.7.1
pywin32==311
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
referencing==0.36.2
regex==2022.10.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rocketchat-API==1.30.0
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentence-transformers==5.1.0
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.36.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.1
starlette==0.47.2
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.14.0
tabulate==0.9.0
tarsafe==0.0.4
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminaltables==3.1.10
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
toolz==1.0.0
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
translator==0.0.9
twilio==8.2.2
typer==0.16.0
typing-inspection==0.4.1
typing-utils==0.1.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wrapt==1.14.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
